Move context to api.go
