+++ gateway/user/api.go
@@ -6,6 +6,7 @@ import (
 	"github.com/getsentry/sentry-go"
 	"github.com/runopsio/hoop/common/log"
 	pb "github.com/runopsio/hoop/common/proto"
+	"go.uber.org/zap"
 
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
@@ -165,4 +166,31 @@ func (a *Handler) UsersGroups(c *gin.Context) {
 	}
 
 	c.PureJSON(http.StatusOK, groups)
+}
+
+// ContextLogger do a best effort to get the context logger,
+// if it fail to retrieve, returns a noop logger
+func ContextLogger(c *gin.Context) *zap.SugaredLogger {
+	obj, _ := c.Get(ContextLoggerKey)
+	if logger := obj.(*zap.SugaredLogger); logger != nil {
+		return logger
+	}
+	return zap.NewNop().Sugar()
+}
+
+// ContextUser do a best effort to get the user context from the request
+// if it fail, it will return an empty one that can be used safely
+func ContextUser(c *gin.Context) *Context {
+	obj, _ := c.Get("context")
+	ctx := obj.(*Context)
+	if ctx == nil {
+		return &Context{Org: &Org{}, User: &User{}}
+	}
+	if ctx.Org == nil {
+		ctx.Org = &Org{}
+	}
+	if ctx.User == nil {
+		ctx.User = &User{}
+	}
+	return ctx
 }
